change of hexes_walked from short->float
Also makes hexes_walked visible via btgetxcodevalue

This will allow us to fix a piloting XP hack in the future.

DB rebuild MAY be required.


git-svn-id: svn://svn.code.sf.net/p/btonline-btech/code/btmux/trunk@1325 e60b412a-5d77-420b-80c2-6cef64185626

+++ btech/src/btech/mech.h
@@ -889,7 +889,7 @@ typedef struct {
     char pilotstatus;		/* damage pilot has taken */
     char terrain;		/* Terrain I am in */
     char elev;			/* Elevation I am at */
-    short hexes_walked;		/* Hexes walked counter */
+    float hexes_walked;		/* Hexes walked counter */
     short facing;		/* 0-359.. */
     short x, y, z;		/* hex quantized x,y,z on the map in MP (hexes) */
     short last_x, last_y;	/* last hex entered */
